Add some additional padding to ensure buttons always float right.
Note in bootstrap v3 (current version we are running right align for columns on a grid method does not exist.
Instead it uses a fake column holder to push the column of interest to the right.
This will not work with an expandable element on the left so instead just use extra padding to ensure the colulm goes on the right.

Refs #298