Adds custom exceptions for handle message
Adds custom exceptions for handle message. If we get into an invalid
state we can probably log and carry on for other runs, since the logic
error could be a missing record or instrument text.

However if we have an implementation error, such as a Value or Type
error then logging and continuing is going to fail for all instruments.
Instead using a custom exception base type we can catch the correct type
in Stomp Client and ignore the other.

Additionally, this prevents some silent return behaviour making it
overall easier to unit test that we throw at the right point, and we can
refactor a return deeper into the stack if required