Actually trace coverage through branches rather than lines
Coveralls by default checks the lines are stepped over, rather than
checking if a branch is entered. E.g:

if False:
    do_something()
bar()

This will show that do_something is covered, despite the fact it's
impossible to run do_something() if bar was called. By setting branch to
true it will correctly show do_something as uncalled.