Revert local sqlite3 usage
Using SQLite3 locally would be a great speed-up for the test execution
(removes remote authentication). However for it to successfully run
fixtures need to be created and inserted before test execution, so that
the default that some tests are expecting are present
(e.g. instruments available: GEM, MARI, etc)

Another stopper is that some tests have MySQL specific calls instead of
using the Django ORM, and thus cannot use SQLite3 backend
without re-writing the test to go through the ORM.