Overrides manual pytest database configuration
Adds a conftest.py that specifies more specific commands that
need to be executed in order to get the DB in an expected state.

In our case we need to migrate and add populate the DB with initial values of
instruments, etc. Currently that information is contained
in the "initial_db_state.json" file - in the future it would be better to
re-use the populate...sql file inside `build/database`