Prevents a plot error from stopping the rendering of the whole page
Moves the plot retrieving logic outside of the main try/except. The
effect is that failing to retrieve a plot for any reason will no longer
prevent the whole page from rendering anything.

Instead if the plot retrieving is failed a `plot_error_message` is passed
to the view and displayed in the place of the plots.