get_icat_instrument_prefix no longer uses VALID_INSTRUMENTS
This was causing an issue that an instrument missing in VALID_INSTRUMENTS
but valid overall, would fail during submission.

The logic for retrieving the instrument's name from ICAT was changed -
get_icat_instrument_prefix now takes a full name and returns the ICAT name.

The behaviour where the function would take a list of instruments and return
a dictionary of {fullname: ICAT name} was never used outside of tests, and has
been removed.

The tests were adjusted to test the new behaviour, and the tests
testing the multiple return behaviour have been removed.

Fixes #842