Now uses the name of the data file to look for plot files
Previously it relied on a manual dictionary mapping between instrument name
and regex that will be used to find the name.

This has been removed in favour of just looking for image files of the same
name as the name of the input data file.

This change shouldn't break plotting for MARI or WISH - and no other instruments
currently use plotting at all, so they can pick up this naming scheme.