Change variable get_or_create to filter. Adds logic to update variables based on tracks_script
Also adds a branch that updates or copies the variable if it's actual contained values are changed.

- Variables will be updated if you've changed a variable in reduce_vars and re-run the same run.
- Variables will be copied if you've changed a variables, but are on a different run.

This accounts for multiple variables with the same value existing. There is a potential futher optimisation, where we move Variable to be related to InstrumentVariable via a foreign key. This allows re-using Variables more easily. However, the duplication might not be severe enough for this to be worth it.