Extract requirements from setup.py into requirements.txt
Move the dev requirements into dev-requirements.txt

The reason for this is that for whatever reason `pip install -e autoreduce` will upgrade packages that autoreduce DOES NOT depend on.
`pip install -r` does not have this behaviour.

This is causing issues on production by thrashing the h5py needed by Mantid.